requirements.txt
requests
python-dotenv
flask
Pillow
pytesseract
